requirements.txt
cryptography
stem
PySocks
